refac: 사용 않는 코드 삭제 및 utils로 파일 이동
