--- test-web/src/Utils/isLogin.js
@@ -1,4 +1,0 @@
-import Cookies from 'js-cookie';
-
-const isLogin = () => !!Cookies.get('token');
-export default isLogin;
