--- test-web/src/components/Base/CustomSnackbar.js
@@ -1,33 +1,0 @@
-import React from 'react';
-import Snackbar from '@mui/material/Snackbar';
-import MuiAlert from '@mui/material/Alert';
-
-const CustomSnackbar = ({ open, message, severity, onClose }) => {
-  const handleSnackbarClose = (event, reason) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    onClose();
-  };
-
-  return (
-    <Snackbar
-      open={open}
-      autoHideDuration={6000}
-      onClose={handleSnackbarClose}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }} // Set the anchorOrigin to top-right
-      style={{ marginTop: '70px' }} // margin for toolbar
-    >
-      <MuiAlert
-        elevation={6}
-        variant="filled"
-        onClose={handleSnackbarClose}
-        severity={severity}
-      >
-        {message}
-      </MuiAlert>
-    </Snackbar>
-  );
-};
-
-export default CustomSnackbar;
