--- test-web/src/components/Base/SelectBtn.js
@@ -1,25 +1,0 @@
-import { Button } from '@mui/material';
-import style from './style/selectbtnstyle';
-
-const SelectedBtn = ({ date }) => {
-  return (
-    <Button variant="contained" value="week" style={style.button}>
-      {date}
-    </Button>
-  );
-};
-
-const UnSelectedBtn = ({ date, handleDr }) => {
-  return (
-    <Button
-      variant="outlined"
-      value="week"
-      style={style.button}
-      onClick={handleDr}
-    >
-      {date}
-    </Button>
-  );
-};
-
-export { SelectedBtn, UnSelectedBtn };
